+++ src/features/ui/Dialog.tsx
@@ -78,7 +78,7 @@ export default function Dialog({
                     as="h3"
                     className={cn(
                       'text-4xl font-semibold text-green-500',
-                      !subtitle && 'mb-4',
+                      !subtitle ? 'mb-4' : 'mb-2',
                     )}
                   >
                     {title}